Packet replay changes
Added:
DebugScreen (F8)
Swedish Translation

Fixed:
Multiconnect ConfigScreen

Changed:
Moved JS/CSS code to write class
packet-log path is a public static final now
