+++ src/main/resources/assets/multiconnect/lang/sv_se.json
@@ -5,6 +5,12 @@
   "multiconnect.config.allowOldUnsignedChat": "Tillåt Gamla Osäkra Meddelanden",
   "multiconnect.config.allowOldUnsignedChat.tooltip": "Tillåter osäkra chattmeddelanden från 1.18.2-servrar eller äldre",
   "multiconnect.config.title": "Multiconnect Konfiguration",
+  "multiconnect.debug.runningHttpServer.stop": "Stoppa Server",
+  "multiconnect.debug.runningHttpServer.title": "Körande HTTP Server",
+  "multiconnect.debug.menu.title": "Multiconnect Debug Skärm",
+  "multiconnect.debug.menu.HttpServer.start" : "Starta HttpServer",
+  "multiconnect.debug.menu.start" : "Starta Replay",
+  "multiconnect.debug.menu.cancel" : "Avbryt",
   "multiconnect.rareBug": "Ett sällsynt fel inträffade just i multiconnect. Ladda upp din spellogg till %s så att det felet kan åtgärdas.",
   "multiconnect.rareBug.link": "Issue Tracker",
   "multiconnect.rareBug.screen": "En sällsynt bugg inträffade just i multiconnect. Vill du ladda upp din spellogg till issue tracker för att hjälpa till att åtgärda problemet?",
